Code review fixes - changed architecture and moved autowired to constructor
