+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/accountmanager/controller/PositionController.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.controller;
+package pl.lodz.p.white.whitetestapp.accountmanager.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,20 +6,23 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import pl.lodz.p.white.whitetestapp.model.Account;
-import pl.lodz.p.white.whitetestapp.service.AccountService;
+import pl.lodz.p.white.whitetestapp.model.Position;
+import pl.lodz.p.white.whitetestapp.accountmanager.service.PositionService;
 
 @Controller
 @RestController
-@RequestMapping("/api/account")
-public class AccountController {
+@RequestMapping("/api/position")
+public class PositionController {
 
-    @Autowired
-    AccountService service;
+    PositionService service;
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    Account get(@PathVariable("id") String id) {
+    Position get(@PathVariable("id") String id){
         return service.getOne(id);
     }
 
+    @Autowired
+    public PositionController(PositionService service) {
+        this.service = service;
+    }
 }
