+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/accountmanager/manager/AccountManager.java
@@ -1,19 +1,23 @@
-package pl.lodz.p.white.whitetestapp.manager;
+package pl.lodz.p.white.whitetestapp.accountmanager.manager;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.lodz.p.white.whitetestapp.model.Account;
 import pl.lodz.p.white.whitetestapp.repository.AccountRepository;
-import pl.lodz.p.white.whitetestapp.service.AccountService;
+import pl.lodz.p.white.whitetestapp.accountmanager.service.AccountService;
 
 @Service
 public class AccountManager implements AccountService {
 
-    @Autowired
     AccountRepository repository;
 
     @Override
     public Account getOne(String id) {
         return repository.getOne(id);
+    }
+
+    @Autowired
+    public AccountManager(AccountRepository repository) {
+        this.repository = repository;
     }
 }
