+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/accountmanager/service/PositionService.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.service;
+package pl.lodz.p.white.whitetestapp.accountmanager.service;
 
 import pl.lodz.p.white.whitetestapp.model.Position;
