+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/controller/AnswerController.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.controller;
+package pl.lodz.p.white.whitetestapp.testmanager.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,19 +6,23 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import pl.lodz.p.white.whitetestapp.model.Position;
-import pl.lodz.p.white.whitetestapp.service.PositionService;
+import pl.lodz.p.white.whitetestapp.model.Answer;
+import pl.lodz.p.white.whitetestapp.testmanager.service.AnswerService;
 
 @Controller
 @RestController
-@RequestMapping("/api/position")
-public class PositionController {
+@RequestMapping("/api/answer")
+public class AnswerController {
 
-    @Autowired
-    PositionService service;
+    AnswerService service;
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    Position get(@PathVariable("id") String id){
+    Answer get(@PathVariable("id") Long id){
         return service.getOne(id);
+    }
+
+    @Autowired
+    public AnswerController(AnswerService service) {
+        this.service = service;
     }
 }
