+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/controller/AnswerToQuestionController.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.controller;
+package pl.lodz.p.white.whitetestapp.testmanager.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -7,18 +7,22 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import pl.lodz.p.white.whitetestapp.model.AnswerToQuestion;
-import pl.lodz.p.white.whitetestapp.service.AnswerToQuestionService;
+import pl.lodz.p.white.whitetestapp.testmanager.service.AnswerToQuestionService;
 
 @Controller
 @RestController
 @RequestMapping("/api/answertoquestion")
 public class AnswerToQuestionController {
 
-    @Autowired
     AnswerToQuestionService service;
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     AnswerToQuestion get(@PathVariable("id") Long id){
         return service.getOne(id);
+    }
+
+    @Autowired
+    public AnswerToQuestionController(AnswerToQuestionService service) {
+        this.service = service;
     }
 }
