+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/controller/QuestionController.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.controller;
+package pl.lodz.p.white.whitetestapp.testmanager.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,19 +6,24 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import pl.lodz.p.white.whitetestapp.model.TestResult;
-import pl.lodz.p.white.whitetestapp.service.TestResultService;
+import pl.lodz.p.white.whitetestapp.model.Question;
+import pl.lodz.p.white.whitetestapp.testmanager.service.QuestionService;
 
 @Controller
 @RestController
-@RequestMapping("/api/testresult")
-public class TestResultController {
+@RequestMapping("/api/question")
+public class QuestionController {
 
-    @Autowired
-    TestResultService service;
+    QuestionService service;
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    TestResult get(@PathVariable("id") Long id){
+    Question get(@PathVariable("id") Long id){
         return service.getOne(id);
     }
+
+    @Autowired
+    public QuestionController(QuestionService service) {
+        this.service = service;
+    }
 }
+
