+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/controller/TestResultController.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.controller;
+package pl.lodz.p.white.whitetestapp.testmanager.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,19 +6,23 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import pl.lodz.p.white.whitetestapp.model.TestTemplate;
-import pl.lodz.p.white.whitetestapp.service.TestTemplateService;
+import pl.lodz.p.white.whitetestapp.model.TestResult;
+import pl.lodz.p.white.whitetestapp.testmanager.service.TestResultService;
 
 @Controller
 @RestController
-@RequestMapping("/api/testtemplate")
-public class TestTemplateController {
+@RequestMapping("/api/testresult")
+public class TestResultController {
 
-    @Autowired
-    TestTemplateService service;
+    TestResultService service;
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    TestTemplate get(@PathVariable("id") Long id){
+    TestResult get(@PathVariable("id") Long id){
         return service.getOne(id);
+    }
+
+    @Autowired
+    public TestResultController(TestResultService service) {
+        this.service = service;
     }
 }
