+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/controller/TestTemplateController.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.controller;
+package pl.lodz.p.white.whitetestapp.testmanager.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,20 +6,23 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import pl.lodz.p.white.whitetestapp.model.Question;
-import pl.lodz.p.white.whitetestapp.service.QuestionService;
+import pl.lodz.p.white.whitetestapp.model.TestTemplate;
+import pl.lodz.p.white.whitetestapp.testmanager.service.TestTemplateService;
 
 @Controller
 @RestController
-@RequestMapping("/api/question")
-public class QuestionController {
+@RequestMapping("/api/testtemplate")
+public class TestTemplateController {
 
-    @Autowired
-    QuestionService service;
+    TestTemplateService service;
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    Question get(@PathVariable("id") Long id){
+    TestTemplate get(@PathVariable("id") Long id){
         return service.getOne(id);
     }
-}
 
+    @Autowired
+    public TestTemplateController(TestTemplateService service) {
+        this.service = service;
+    }
+}
