+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/manager/AnswerManager.java
@@ -1,19 +1,23 @@
-package pl.lodz.p.white.whitetestapp.manager;
+package pl.lodz.p.white.whitetestapp.testmanager.manager;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.lodz.p.white.whitetestapp.model.Answer;
 import pl.lodz.p.white.whitetestapp.repository.AnswerRepository;
-import pl.lodz.p.white.whitetestapp.service.AnswerService;
+import pl.lodz.p.white.whitetestapp.testmanager.service.AnswerService;
 
 @Service
 public class AnswerManager implements AnswerService {
 
-    @Autowired
     AnswerRepository repository;
 
     @Override
     public Answer getOne(Long id) {
         return repository.getOne(id);
+    }
+
+    @Autowired
+    public AnswerManager(AnswerRepository repository) {
+        this.repository = repository;
     }
 }
