+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/manager/AnswerToQuestionManager.java
@@ -1,19 +1,23 @@
-package pl.lodz.p.white.whitetestapp.manager;
+package pl.lodz.p.white.whitetestapp.testmanager.manager;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.lodz.p.white.whitetestapp.model.AnswerToQuestion;
 import pl.lodz.p.white.whitetestapp.repository.AnswerToQuestionRepository;
-import pl.lodz.p.white.whitetestapp.service.AnswerToQuestionService;
+import pl.lodz.p.white.whitetestapp.testmanager.service.AnswerToQuestionService;
 
 @Service
 public class AnswerToQuestionManager implements AnswerToQuestionService {
 
-    @Autowired
     AnswerToQuestionRepository repository;
 
     @Override
     public AnswerToQuestion getOne(Long id) {
         return repository.getOne(id);
+    }
+
+    @Autowired
+    public AnswerToQuestionManager(AnswerToQuestionRepository repository) {
+        this.repository = repository;
     }
 }
