+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/manager/QuestionManager.java
@@ -1,19 +1,23 @@
-package pl.lodz.p.white.whitetestapp.manager;
+package pl.lodz.p.white.whitetestapp.testmanager.manager;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.lodz.p.white.whitetestapp.model.Question;
 import pl.lodz.p.white.whitetestapp.repository.QuestionRepository;
-import pl.lodz.p.white.whitetestapp.service.QuestionService;
+import pl.lodz.p.white.whitetestapp.testmanager.service.QuestionService;
 
 @Service
 public class QuestionManager implements QuestionService {
 
-    @Autowired
     QuestionRepository repository;
 
     @Override
     public Question getOne(Long id) {
         return repository.getOne(id);
+    }
+
+    @Autowired
+    public QuestionManager(QuestionRepository repository) {
+        this.repository = repository;
     }
 }
