+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/manager/TestTemplateManager.java
@@ -1,19 +1,23 @@
-package pl.lodz.p.white.whitetestapp.manager;
+package pl.lodz.p.white.whitetestapp.testmanager.manager;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.lodz.p.white.whitetestapp.model.TestTemplate;
 import pl.lodz.p.white.whitetestapp.repository.TestTemplateRepository;
-import pl.lodz.p.white.whitetestapp.service.TestTemplateService;
+import pl.lodz.p.white.whitetestapp.testmanager.service.TestTemplateService;
 
 @Service
 public class TestTemplateManager implements TestTemplateService {
 
-    @Autowired
     TestTemplateRepository repository;
 
     @Override
     public TestTemplate getOne(Long id) {
         return repository.getOne(id);
+    }
+
+    @Autowired
+    public TestTemplateManager(TestTemplateRepository repository) {
+        this.repository = repository;
     }
 }
