+++ white-test-app-backend/src/main/java/pl/lodz/p/white/whitetestapp/testmanager/service/QuestionService.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.white.whitetestapp.service;
+package pl.lodz.p.white.whitetestapp.testmanager.service;
 
 import pl.lodz.p.white.whitetestapp.model.Question;
